refactor(back): rename class
