+++ backend/moneyFUS/urls.py
@@ -3,5 +3,5 @@ from . import views
 
 urlpatterns = [
     path("user/login", views.user.login.LoginView.as_view()),
-    path("user", views.user.register.RegisterView.as_view())
+    path("user", views.user.index.UserView.as_view()),
 ]
